Add fake_environment class to ease unit tests of modules
With the assumption the most modules for the sonic collection will take some parameters
and modify the SONiC configuration database using the API from swsscommon the class can
be used to setup an environment where the modules can run without the swsscommon module.

It does this by changing changing python module internals and might not work when
running tests in parallel.

Signed-off-by: Erik Larsson <who+github@cnackers.org>